requirements/base.txt
Django>=3.1.0,<3.2.0
wagtail>=2.11.0,<2.12.0
wagtail-bifrost>=3.0.0
psycopg2-binary==2.8.4
wagtail-django-recaptcha==1.0
django-pattern-library==0.3.0
wagtailfontawesome==1.1.4
graphene-django-optimizer==0.6.2
django-cors-headers==3.4.0
django-filter==2.4.0
wagtail-generic-chooser==0.1.1
redis==3.5.3
django-redis==4.12.1
channels-redis==3.2.0
graphql-ws==0.3.1
wagtailmedia==0.5.0
factory-boy==2.12.0
wagtail-factories==2.0.0
wagtail-headless-preview==0.1.4